requirements.txt
python-telegram-bot
twitter-api-client